requirements.txt
python-telegram-bot==20.6
pandas==2.1.2
numpy==1.26.1
psycopg2-binary==2.9.9
Flask==3.0.0
matplotlib==3.8.1
seaborn==0.13.0
psutil==5.9.6
python-dotenv==1.0.0
Flask
matplotlib
numpy
pandas
psutil
psycopg2-binary
python-dotenv
python-telegram-bot
requests
seaborn
telegram
yfinance==0.2.55
Flask
matplotlib
numpy
pandas
psutil
psycopg2-binary
python-dotenv
python-telegram-bot
requests
seaborn
telegram
yfinance
pillow
matplotlib
numpy
pillow
python-telegram-bot
telegram
Flask
matplotlib
numpy
pandas
pillow
psutil
psycopg2-binary
python-dotenv
requests
seaborn
yfinance
flask
matplotlib
numpy
pandas
pillow
psutil
psycopg2-binary
python-dotenv
python-telegram-bot
requests
seaborn
telegram
yfinance
pytz
flask
matplotlib
numpy
pandas
pillow
psutil
psycopg2-binary
python-dotenv
python-telegram-bot==20.6
requests
seaborn
telegram
yfinance
gunicorn
